# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeSentinel AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReviewStatus = typeof ReviewStatus[keyof typeof ReviewStatus];


export const ReviewStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Review {
  id: number;
  title: string;
  /** @nullable */
  prUrl?: string | null;
  status: ReviewStatus;
  totalFindings: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  /** @nullable */
  language?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ReviewInput {
  title: string;
  prUrl?: string;
  diffContent?: string;
  language?: string;
}

export type FindingCategory = typeof FindingCategory[keyof typeof FindingCategory];


export const FindingCategory = {
  security: 'security',
  performance: 'performance',
  code_quality: 'code_quality',
  testing: 'testing',
  maintainability: 'maintainability',
} as const;

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Finding {
  id: number;
  reviewId: number;
  category: FindingCategory;
  severity: FindingSeverity;
  title: string;
  description: string;
  suggestion: string;
  /** @nullable */
  filePath?: string | null;
  /** @nullable */
  lineNumber?: number | null;
  /** @nullable */
  codeSnippet?: string | null;
  createdAt: string;
}

export type ReviewWithFindingsStatus = typeof ReviewWithFindingsStatus[keyof typeof ReviewWithFindingsStatus];


export const ReviewWithFindingsStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ReviewWithFindings {
  id: number;
  title: string;
  /** @nullable */
  prUrl?: string | null;
  status: ReviewWithFindingsStatus;
  /** @nullable */
  diffContent?: string | null;
  totalFindings: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  /** @nullable */
  language?: string | null;
  createdAt: string;
  updatedAt: string;
  findings: Finding[];
}

export type ReviewStatsFindingsByCategory = {[key: string]: number};

export interface ReviewStats {
  totalReviews: number;
  totalFindings: number;
  criticalFindings: number;
  highFindings: number;
  mediumFindings: number;
  lowFindings: number;
  findingsByCategory: ReviewStatsFindingsByCategory;
  recentActivity: Review[];
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: number;
  reviewId: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface ChatMessageInput {
  content: string;
}

export interface ApiError {
  error: string;
}