# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeSentinel AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingCategory } from './findingCategory';
import type { FindingSeverity } from './findingSeverity';

export interface Finding {
  id: number;
  reviewId: number;
  category: FindingCategory;
  severity: FindingSeverity;
  title: string;
  description: string;
  suggestion: string;
  /** @nullable */
  filePath?: string | null;
  /** @nullable */
  lineNumber?: number | null;
  /** @nullable */
  codeSnippet?: string | null;
  createdAt: Date;
}