# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeSentinel AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewStatus } from './reviewStatus';

export interface Review {
  id: number;
  title: string;
  /** @nullable */
  prUrl?: string | null;
  status: ReviewStatus;
  totalFindings: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  /** @nullable */
  language?: string | null;
  createdAt: Date;
  updatedAt: Date;
}