# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeSentinel AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewInput {
  title: string;
  prUrl?: string;
  diffContent?: string;
  language?: string;
}