# lib/api-zod/src/generated/types/reviewWithFindings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeSentinel AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Finding } from './finding';
import type { ReviewWithFindingsStatus } from './reviewWithFindingsStatus';

export interface ReviewWithFindings {
  id: number;
  title: string;
  /** @nullable */
  prUrl?: string | null;
  status: ReviewWithFindingsStatus;
  /** @nullable */
  diffContent?: string | null;
  totalFindings: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  /** @nullable */
  language?: string | null;
  createdAt: Date;
  updatedAt: Date;
  findings: Finding[];
}